LinkedList.java
Implementation of LinkedList from scratch
